+++ ChangeLog.md
@@ -10,6 +10,7 @@
 - file address abbrs of *.abbr for export
 - file address abbrs of *.abbr or dir path for import
 - flag allowing to set upper case in words
+- flag that takes into account when assembling N-grams, only those tokens that match words
 
 ## [3.2.6](https://github.com/anyks/alm/archive/v3.2.6.tar.gz) Other optimization